requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
youtube-transcript-api==1.0.2
google-generativeai==0.3.0
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.12.0
sentence-transformers==2.2.2
torch==2.8.0
transformers==4.57.1
nltk==3.8.1
openai==1.3.5
pydantic==2.4.2
Werkzeug==2.3.7
pydub==0.25.1
yt_dlp
google-search-results==2.4.2
